Use console heartbeats to detect when all consoles have disconnected and provide a warning
Arguably this should be a configurable command, but first another matter must be sorted out:
this dredged up the fact that commands mostly need to run on the foreground thread, even if
they come off the queue in the background thread. For now, the queue was moved to the
foreground; in the future, this will need a more nuanced solution.