Disambiguate ScopedConfig logging -- show specific key to humans and include both in JSON
The biggest problem with the ScopedConfig class was that it made
logging unclear. A simple solution, like adding an extra log line
when a generic config is used, would mean doubling the amount of
logging in many applications. It would also greatly reduce clarity.

Instead, recognizing that humans and machines have different needs
and that debugging a hard to find config is a rare scenario, this
change always shows the specific config to humans, even when a
generic one is read. The JSON includes both for machine analysis.