Fix config source priorities without duplicating primary config
Previously, the primary config source was loaded a second time, after loading
any includes it specified. This was was required so the primary config could
override the values in those includes; however, the duplication was ugly.
Since the config sources had to be saved separately to set their logs later,
it was simple enough to clear them out and reinsert them all in the correct
order.