Add option to dry run, executing no update cycles and shutting down immediately
Dry runs are a great way to test configurations before deploying them. Having
the option to do that directly in the application allows all of the usual
config validations to be run, while still ensuring safety by not executing any
updates.

The flag can be set with -d on the command line, or via the Application.DryRun
config:

	> build/bin/HelloWorld -d | scripts/FormatLog.py
	[00000] [1656438727.859776000]     INFO - <Config> Loading src/HelloWorld/TestRealTime.ini...
	...
	[00010] [1656438727.861689000]   CONFIG - <Config> [Application.DryRun] = false [DEFAULT] -- Indicates whether the application should do a dry run (no update cycles; immediate shutdown)
	[00012] [1656438727.861690000]     INFO - <Context> Executing dry run (no update cycles)