Add functional test harness and basic tests
For full test coverage, unit tests are not enough: end-to-end testing by
running the final binaries can reveal problems with integration that would be
invisible to traditional unit testing. Even elaborate integration testing is
a poor substitute for the real thing.

This change introduces a functional test harness via the python `unittest`
module. This handles all of the basic requirements of a test framework, and
allows the test logic to be written in Python. Included are several examples
of simple tests. For example, the version binary is run to ensure it produces
a valid git hash and nothing more. In the future, this test suite will be
expanded to cover many facets of the framework.