Functional tests for segfault command
The segfault command can be useful in the right context and extremely dangerous
in the wrong one. The command must only be available in simulation; there is no
reason to segfault a production application. This is now verified with two new
functional tests:
1) The "happy path": Simulation is true, the command runs, and the correct
   output and error code are returned.
2) The "unhappy path": Simulation is false, the command does not exist, and the
   corresponding error is logged.

During the testing of this change, it became clear that the post-segfault waits
were too long at 5 seconds. They have been reduced to 1 second.